refactor to split lib into common, ascii_protocol and meta_protocol files
