+++ src/lib.rs
@@ -5,16 +5,19 @@ use bytes::BytesMut;
 use fxhash::FxHashMap;
 use tokio::io::{AsyncBufReadExt, AsyncReadExt, AsyncWriteExt};
 
+mod ascii_protocol;
+
 mod connection;
 use self::connection::Connection;
 
 mod error;
 pub use self::error::Error;
 
+mod meta_protocol;
+
 mod parser;
 use self::parser::{
-    parse_ascii_metadump_response, parse_ascii_response, parse_ascii_stats_response,
-    parse_meta_response, Response,
+    parse_ascii_metadump_response, parse_ascii_response, parse_ascii_stats_response, Response,
 };
 pub use self::parser::{ErrorKind, KeyMetadata, MetadumpResponse, StatsResponse, Status, MetaValue, Value, MemcachedValue};
 
@@ -131,33 +134,6 @@ impl Client {
         self.drive_receive(parse_ascii_stats_response).await
     }
 
-    /// Gets the given key.
-    ///
-    /// If the key is found, `Some(Value)` is returned, describing the metadata and data of the key.
-    ///
-    /// Otherwise, [`Error`] is returned.
-    pub async fn get<K: AsRef<[u8]>>(&mut self, key: K) -> Result<Option<MemcachedValue>, Error> {
-        self.conn
-            .write_all(&[b"get ", key.as_ref(), b"\r\n"].concat())
-            .await?;
-        self.conn.flush().await?;
-
-        match self.get_read_write_response().await? {
-            Response::Status(Status::NotFound) => Ok(None),
-            Response::Status(s) => Err(s.into()),
-            Response::Data(d) => d
-                .map(|mut items| {
-                    if items.len() != 1 {
-                        Err(Status::Error(ErrorKind::Protocol(None)).into())
-                    } else {
-                        Ok(items.remove(0))
-                    }
-                })
-                .transpose(),
-            _ => Err(Error::Protocol(Status::Error(ErrorKind::Protocol(None)))),
-        }
-    }
-
     /// Gets the given key with additional metadata.
     ///
     /// If the key is found, `Some(Value)` is returned, describing the metadata and data of the key.
@@ -168,223 +144,7 @@ impl Client {
     /// - h: return whether item has been hit before as a 0 or 1
     /// - l: return time since item was last accessed in seconds
     /// - t: return item TTL remaining in seconds (-1 for unlimited)
-    pub async fn meta_get<K: AsRef<[u8]>>(
-        &mut self,
-        key: K,
-        meta_flags: &[char],
-    ) -> Result<Option<MetaValue>, Error> {
-        let mut command = Vec::with_capacity(64);
-        command.extend_from_slice(b"mg ");
-        command.extend_from_slice(key.as_ref());
-        // TODO: Do we want to always add the v flag? so all gets return the value? or we want clients to know to request it?
-        command.extend_from_slice(b" v");
-        if !meta_flags.is_empty() {
-            for flag in meta_flags {
-                command.push(b' ');
-                command.push(*flag as u8);
-            }
-        }
-        command.extend_from_slice(b"\r\n");
-        //println!("command: {:?}", String::from_utf8_lossy(&command));
-        self.conn.write_all(&command).await?;
-        self.conn.flush().await?;
 
-        match self.drive_receive(parse_meta_response).await? {
-            Response::Status(Status::NotFound) => Ok(None),
-            Response::Status(s) => Err(s.into()),
-            Response::Data(d) => d
-                .map(|mut items| {
-                    if items.len() != 1 {
-                        Err(Status::Error(ErrorKind::Protocol(None)).into())
-                    } else {
-                        let item = items.remove(0);
-                        match item {
-                            MemcachedValue::MetaValue(mut item) => {
-                                item.key = Some(key.as_ref().to_vec());
-                                Ok(item)
-                            }
-                            _ => panic!("Expected MemcachedValue::MetaValue, got something else"),
-                        }
-                    }
-                })
-                .transpose(),
-            _ => Err(Error::Protocol(Status::Error(ErrorKind::Protocol(None)))),
-        }
-    }
-
-    /// Gets the given keys.
-    ///
-    /// If any of the keys are found, a vector of [`Value`] will be returned, where [`Value`]
-    /// describes the metadata and data of the key.
-    ///
-    /// Otherwise, [`Error`] is returned.
-    pub async fn get_multi<I, K>(&mut self, keys: I) -> Result<Vec<MemcachedValue>, Error>
-    where
-        I: IntoIterator<Item = K>,
-        K: AsRef<[u8]>,
-    {
-        self.conn.write_all(b"get ").await?;
-        for key in keys {
-            self.conn.write_all(key.as_ref()).await?;
-            self.conn.write_all(b" ").await?;
-        }
-        self.conn.write_all(b"\r\n").await?;
-        self.conn.flush().await?;
-
-        match self.get_read_write_response().await? {
-            Response::Status(s) => Err(s.into()),
-            Response::Data(d) => d.ok_or(Status::NotFound.into()),
-            _ => Err(Status::Error(ErrorKind::Protocol(None)).into()),
-        }
-    }
-
-    /// Gets the given keys.
-    ///
-    /// Deprecated: This is now an alias for `get_multi`, and  will be removed in the future.
-    #[deprecated(
-        since = "0.4.0",
-        note = "This is now an alias for `get_multi`, and will be removed in the future."
-    )]
-    pub async fn get_many<I, K>(&mut self, keys: I) -> Result<Vec<MemcachedValue>, Error>
-    where
-        I: IntoIterator<Item = K>,
-        K: AsRef<[u8]>,
-    {
-        self.get_multi(keys).await
-    }
-
-    /// Sets the given key.
-    ///
-    /// If `ttl` or `flags` are not specified, they will default to 0.  If the value is set
-    /// successfully, `()` is returned, otherwise [`Error`] is returned.
-    pub async fn set<K, V>(
-        &mut self,
-        key: K,
-        value: V,
-        ttl: Option<i64>,
-        flags: Option<u32>,
-    ) -> Result<(), Error>
-    where
-        K: AsRef<[u8]>,
-        V: AsMemcachedValue,
-    {
-        let kr = key.as_ref();
-        let vr = value.as_bytes();
-
-        self.conn.write_all(b"set ").await?;
-        self.conn.write_all(kr).await?;
-
-        let flags = flags.unwrap_or(0).to_string();
-        self.conn.write_all(b" ").await?;
-        self.conn.write_all(flags.as_ref()).await?;
-
-        let ttl = ttl.unwrap_or(0).to_string();
-        self.conn.write_all(b" ").await?;
-        self.conn.write_all(ttl.as_ref()).await?;
-
-        let vlen = vr.len().to_string();
-        self.conn.write_all(b" ").await?;
-        self.conn.write_all(vlen.as_ref()).await?;
-        self.conn.write_all(b"\r\n").await?;
-
-        self.conn.write_all(vr.as_ref()).await?;
-        self.conn.write_all(b"\r\n").await?;
-
-        self.conn.flush().await?;
-
-        match self.get_read_write_response().await? {
-            Response::Status(Status::Stored) => Ok(()),
-            Response::Status(s) => Err(s.into()),
-            _ => Err(Status::Error(ErrorKind::Protocol(None)).into()),
-        }
-    }
-
-    /// Sets multiple keys and values through pipelined commands.
-    ///
-    /// If `ttl` or `flags` are not specified, they will default to 0. The same values for `ttl` and `flags` will be applied to each key.
-    /// Returns a result with a HashMap of keys mapped to the result of the set operation, or an error.
-    pub async fn set_multi<'a, K, V>(
-        &mut self,
-        kv: &'a [(K, V)],
-        ttl: Option<i64>,
-        flags: Option<u32>,
-    ) -> Result<FxHashMap<&'a K, Result<(), Error>>, Error>
-    where
-        K: AsRef<[u8]> + Eq + std::hash::Hash + std::fmt::Debug,
-        V: AsMemcachedValue,
-    {
-        for (key, value) in kv {
-            let kr = key.as_ref();
-            let vr = value.as_bytes();
-
-            self.conn.write_all(b"set ").await?;
-            self.conn.write_all(kr).await?;
-
-            let flags = flags.unwrap_or(0).to_string();
-            self.conn.write_all(b" ").await?;
-            self.conn.write_all(flags.as_ref()).await?;
-
-            let ttl = ttl.unwrap_or(0).to_string();
-            self.conn.write_all(b" ").await?;
-            self.conn.write_all(ttl.as_ref()).await?;
-
-            let vlen = vr.len().to_string();
-            self.conn.write_all(b" ").await?;
-            self.conn.write_all(vlen.as_ref()).await?;
-            self.conn.write_all(b"\r\n").await?;
-
-            self.conn.write_all(vr.as_ref()).await?;
-            self.conn.write_all(b"\r\n").await?;
-        }
-        self.conn.flush().await?;
-
-        let results = self.map_set_multi_responses(kv).await?;
-
-        Ok(results)
-    }
-
-    /// Add a key. If the value exists, Err(Protocol(NotStored)) is returned.
-    pub async fn add<K, V>(
-        &mut self,
-        key: K,
-        value: V,
-        ttl: Option<i64>,
-        flags: Option<u32>,
-    ) -> Result<(), Error>
-    where
-        K: AsRef<[u8]>,
-        V: AsMemcachedValue,
-    {
-        let kr = key.as_ref();
-        let vr = value.as_bytes();
-
-        self.conn.write_all(b"add ").await?;
-        self.conn.write_all(kr).await?;
-
-        let flags = flags.unwrap_or(0).to_string();
-        self.conn.write_all(b" ").await?;
-        self.conn.write_all(flags.as_ref()).await?;
-
-        let ttl = ttl.unwrap_or(0).to_string();
-        self.conn.write_all(b" ").await?;
-        self.conn.write_all(ttl.as_ref()).await?;
-
-        let vlen = vr.len().to_string();
-        self.conn.write_all(b" ").await?;
-        self.conn.write_all(vlen.as_ref()).await?;
-        self.conn.write_all(b"\r\n").await?;
-
-        self.conn.write_all(vr.as_ref()).await?;
-        self.conn.write_all(b"\r\n").await?;
-
-        self.conn.flush().await?;
-
-        match self.get_read_write_response().await? {
-            Response::Status(Status::Stored) => Ok(()),
-            Response::Status(s) => Err(s.into()),
-            _ => Err(Status::Error(ErrorKind::Protocol(None)).into()),
-        }
-    }
 
     /// Sets the given key.
     ///
@@ -402,268 +162,7 @@ impl Client {
     //
     // - <flags> are a set of single character codes ended with a space or newline.
     //   flags may have strings after the initial character.
-    pub async fn meta_set<K, V>(
-        &mut self,
-        key: K,
-        value: V,
-        ttl: Option<i64>,
-        meta_flags: FxHashMap<&[char], String>,
-    ) -> Result<Option<MetaValue>, Error>
-    where
-        K: AsRef<[u8]>,
-        V: AsMemcachedValue,
-    {
-        let mut command = Vec::with_capacity(64);
-        let kr = key.as_ref();
-        let vr = value.as_bytes();
 
-        command.extend_from_slice(b"ms ");
-        command.extend_from_slice(kr);
-        command.extend_from_slice(b" ");
-        let vlen = vr.len().to_string();
-        command.extend_from_slice(vlen.as_ref());
-
-        if !meta_flags.is_empty() {
-            for (flag, flag_value) in meta_flags {
-                command.push(b' ');
-                command.extend(flag.iter().map(|&c| c as u8));
-                if !flag_value.is_empty() {
-                    command.extend_from_slice(flag_value.as_bytes());
-                }
-            }
-        }
-
-        // TTL is just another flag, but we encourage passing it to the function for clarity
-        let ttl = ttl.unwrap_or(0).to_string();
-        command.extend_from_slice(b" ");
-        command.extend_from_slice(b"T");
-        command.extend_from_slice(ttl.as_ref());
-        command.extend_from_slice(b"\r\n");
-        self.conn.write_all(&command).await?;
-        println!("command: {:?}", String::from_utf8_lossy(&command));
-        self.conn.write_all(vr.as_ref()).await?;
-        self.conn.write_all(b"\r\n").await?;
-        self.conn.flush().await?;
-
-        match self.drive_receive(parse_meta_response).await? {
-            Response::Status(s) => Err(s.into()),
-            Response::Data(d) => d
-                .map(|mut items| {
-                    if items.len() != 1 {
-                        Err(Status::Error(ErrorKind::Protocol(None)).into())
-                    } else {
-                        let item = items.remove(0);
-                        match item {
-                            MemcachedValue::MetaValue(mut item) => {
-                                item.key = Some(key.as_ref().to_vec());
-                                Ok(item)
-                            }
-                            _ => panic!("Expected MemcachedValue::MetaValue, got something else"),
-                        }
-                    }
-                })
-                .transpose(),
-            _ => Err(Error::Protocol(Status::Error(ErrorKind::Protocol(None)))),
-        }
-    }
-
-    /// Attempts to add multiple keys and values through pipelined commands.
-    ///
-    /// If `ttl` or `flags` are not specified, they will default to 0. The same values for `ttl` and `flags` will be applied to each key.
-    /// Returns a result with a HashMap of keys mapped to the result of the add operation, or an error.
-    pub async fn add_multi<'a, K, V>(
-        &mut self,
-        kv: &'a [(K, V)],
-        ttl: Option<i64>,
-        flags: Option<u32>,
-    ) -> Result<FxHashMap<&'a K, Result<(), Error>>, Error>
-    where
-        K: AsRef<[u8]> + Eq + std::hash::Hash + std::fmt::Debug,
-        V: AsMemcachedValue,
-    {
-        for (key, value) in kv {
-            let kr = key.as_ref();
-            let vr = value.as_bytes();
-
-            self.conn.write_all(b"add ").await?;
-            self.conn.write_all(kr).await?;
-
-            let flags = flags.unwrap_or(0).to_string();
-            self.conn.write_all(b" ").await?;
-            self.conn.write_all(flags.as_ref()).await?;
-
-            let ttl = ttl.unwrap_or(0).to_string();
-            self.conn.write_all(b" ").await?;
-            self.conn.write_all(ttl.as_ref()).await?;
-
-            let vlen = vr.len().to_string();
-            self.conn.write_all(b" ").await?;
-            self.conn.write_all(vlen.as_ref()).await?;
-            self.conn.write_all(b"\r\n").await?;
-
-            self.conn.write_all(vr.as_ref()).await?;
-            self.conn.write_all(b"\r\n").await?;
-        }
-        self.conn.flush().await?;
-
-        let results = self.map_set_multi_responses(kv).await?;
-
-        Ok(results)
-    }
-
-    /// Delete a key but don't wait for a reply.
-    pub async fn delete_no_reply<K>(&mut self, key: K) -> Result<(), Error>
-    where
-        K: AsRef<[u8]>,
-    {
-        let kr = key.as_ref();
-
-        self.conn
-            .write_all(&[b"delete ", kr, b" noreply\r\n"].concat())
-            .await?;
-        self.conn.flush().await?;
-        Ok(())
-    }
-
-    /// Delete a key and wait for a reply
-    pub async fn delete<K>(&mut self, key: K) -> Result<(), Error>
-    where
-        K: AsRef<[u8]>,
-    {
-        let kr = key.as_ref();
-
-        self.conn
-            .write_all(&[b"delete ", kr, b"\r\n"].concat())
-            .await?;
-        self.conn.flush().await?;
-
-        match self.get_read_write_response().await? {
-            Response::Status(Status::Deleted) => Ok(()),
-            Response::Status(s) => Err(s.into()),
-            _ => Err(Status::Error(ErrorKind::Protocol(None)).into()),
-        }
-    }
-
-    /// Delete multiple keys
-    pub async fn delete_multi_no_reply<K>(&mut self, keys: &[K]) -> Result<(), Error>
-    where
-        K: AsRef<[u8]>,
-    {
-        for key in keys {
-            self.conn.write_all(b"delete ").await?;
-            self.conn.write_all(key.as_ref()).await?;
-            self.conn.write_all(b" noreply\r\n").await?;
-        }
-        self.conn.flush().await?;
-
-        Ok(())
-    }
-
-    /// Increments the given key by the specified amount.
-    /// Can overflow from the max value of u64 (18446744073709551615) -> 0.
-    /// If the key does not exist, the server will return a KeyNotFound error.
-    /// If the key exists but the value is non-numeric, the server will return a ClientError.
-    pub async fn increment<K>(&mut self, key: K, amount: u64) -> Result<u64, Error>
-    where
-        K: AsRef<[u8]>,
-    {
-        self.conn
-            .write_all(
-                &[
-                    b"incr ",
-                    key.as_ref(),
-                    b" ",
-                    amount.to_string().as_bytes(),
-                    b"\r\n",
-                ]
-                .concat(),
-            )
-            .await?;
-        self.conn.flush().await?;
-
-        match self.get_read_write_response().await? {
-            Response::Status(s) => Err(s.into()),
-            Response::IncrDecr(amount) => Ok(amount),
-            _ => Err(Status::Error(ErrorKind::Protocol(None)).into()),
-        }
-    }
-
-    /// Increments the given key by the specified amount with no reply from the server.
-    /// Can overflow from the max value of u64 (18446744073709551615) -> 0.
-    /// Always returns () for a complete request, will not return any indication of success or failure.
-    pub async fn increment_no_reply<K>(&mut self, key: K, amount: u64) -> Result<(), Error>
-    where
-        K: AsRef<[u8]>,
-    {
-        self.conn
-            .write_all(
-                &[
-                    b"incr ",
-                    key.as_ref(),
-                    b" ",
-                    amount.to_string().as_bytes(),
-                    b" noreply\r\n",
-                ]
-                .concat(),
-            )
-            .await?;
-        self.conn.flush().await?;
-
-        Ok(())
-    }
-
-    /// Decrements the given key by the specified amount.
-    /// Will not decrement the counter below 0.
-    /// If the key does not exist, the server will return a KeyNotFound error.
-    /// If the key exists but the value is non-numeric, the server will return a ClientError.
-    pub async fn decrement<K>(&mut self, key: K, amount: u64) -> Result<u64, Error>
-    where
-        K: AsRef<[u8]>,
-    {
-        self.conn
-            .write_all(
-                &[
-                    b"decr ",
-                    key.as_ref(),
-                    b" ",
-                    amount.to_string().as_bytes(),
-                    b"\r\n",
-                ]
-                .concat(),
-            )
-            .await?;
-        self.conn.flush().await?;
-
-        match self.get_read_write_response().await? {
-            Response::Status(s) => Err(s.into()),
-            Response::IncrDecr(amount) => Ok(amount),
-            _ => Err(Status::Error(ErrorKind::Protocol(None)).into()),
-        }
-    }
-
-    /// Decrements the given key by the specified amount with no reply from the server.
-    /// Will not decrement the counter below 0.
-    /// Always returns () for a complete request, will not return any indication of success or failure.
-    pub async fn decrement_no_reply<K>(&mut self, key: K, amount: u64) -> Result<(), Error>
-    where
-        K: AsRef<[u8]>,
-    {
-        self.conn
-            .write_all(
-                &[
-                    b"decr ",
-                    key.as_ref(),
-                    b" ",
-                    amount.to_string().as_bytes(),
-                    b" noreply\r\n",
-                ]
-                .concat(),
-            )
-            .await?;
-        self.conn.flush().await?;
-
-        Ok(())
-    }
 
     /// Gets the version of the server.
     ///
@@ -690,7 +189,6 @@ impl Client {
     }
 
     /// Dumps all keys from the server.
-    ///
     /// This operation scans all slab classes from tail to head, in a non-blocking fashion.  Thus,
     /// not all items will be found as new items could be inserted or deleted while the crawler is
     /// still running.
