+++ src/parser/mod.rs
@@ -9,12 +9,10 @@ use nom::{
 };
 use std::fmt;
 
-mod ascii;
+pub mod ascii;
 pub use ascii::{parse_ascii_metadump_response, parse_ascii_response, parse_ascii_stats_response};
 
-mod meta;
-pub use meta::parse_meta_response;
-
+pub mod meta;
 /// A value from memcached generated by an ASCII protocol response.
 #[derive(Clone, Debug, PartialEq)]
 pub struct Value {
